Fix buffer overflow in buildboardinfodialog()
The board size message can be up to 22 characters long, which causes a
buffer overflow when sprintf() writes into buffer[20]. When built on
Ubuntu 16.04 LTS with default settings, glibc detects this at runtime
and terminates KevEdit with "buffer overflow detected".

This commit increases buffer's size to 24 bytes, just to have a round
number.