Use CTRL as a modifier for moving between boards
This is another hack I have been using for a long time. CTRL + direction to move between linked boards and CTRL+PgUp / CTRL+PgDn to move the same number of boards as PgUp/PgDn moves you in the 'B'oard selector. It's intuitive and it makes checking the connections between rooms really easy.