First attempt at a new SDL audio callback
Rather than allocating and reallocating a monster buffer of the entire
audio stream, generate it on demand from a smaller buffer of notes.

This still needs:
- callback emits an event to tell the player it needs more notes
- wait until buffer is drained before closing audio

It would also be nice to make a less harsh waveform.  It should be
relatively simple to add polybleps on the samples surrounding the
transition.