Make CTRL+A available in line_editor for misc character input.
This should allow extended chars in things like board names.
It also implements textBlockBlit and a backup buffer storing
the contents of the screen behind the popup so they can be restored
afterwards.

Tested on SDL, implemented but not tested on DOS, not implemented on
stdio (will crash). I'll fix that in a later commit.