Make character selection starting point depend on last selected char.
This might be a more convenient option: the selector's starting point
now depends on where it was placed last time the selector was used
in that line editor, instead of the last character typed.

Say for instance the user wants to put a gradient in a board title,
then type in the title name, then another gradient. With this
modification, the selector will start off at the gradient characters
rather than the last typed letter, which might be less of a hassle.
On the other hand, it could also be less intuitive.