Remove old pasteblock function.
We might argue that the new pasting function is so complex that it
shouldn't be used until the user has decided where to place the
selection. But if that's a problem, a simple "|| preview_mode" on
the "If there are no objects to add" if statement should do the job
much more easily.